seg-service/requirements.txt
fastapi
uvicorn
onnx
onnxruntime
pillow
opencv-python-headless